feat!: yargs is now ESM first
